+++ tendermint/src/abci/request.rs
@@ -55,62 +55,3 @@ pub use process_proposal::ProcessProposal;
 pub use query::Query;
 pub use set_option::SetOption;
 pub use verify_vote_extension::VerifyVoteExtension;
-
-/// The consensus category of ABCI requests.
-#[allow(clippy::large_enum_variant)]
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub enum ConsensusRequest {
-    #[doc = include_str!("doc/request-initchain.md")]
-    InitChain(InitChain),
-    #[doc = include_str!("doc/request-prepareproposal.md")]
-    PrepareProposal(PrepareProposal),
-    #[doc = include_str!("doc/request-processproposal.md")]
-    ProcessProposal(ProcessProposal),
-    #[doc = include_str!("doc/request-beginblock.md")]
-    BeginBlock(BeginBlock),
-    #[doc = include_str!("doc/request-delivertx.md")]
-    DeliverTx(DeliverTx),
-    #[doc = include_str!("doc/request-endblock.md")]
-    EndBlock(EndBlock),
-    #[doc = include_str!("doc/request-commit.md")]
-    Commit,
-    #[doc = include_str!("doc/request-extendvote.md")]
-    ExtendVote(ExtendVote),
-    #[doc = include_str!("doc/request-verifyvoteextension.md")]
-    VerifyVoteExtension(VerifyVoteExtension),
-    #[doc = include_str!("doc/request-finalizeblock.md")]
-    FinalizeBlock(FinalizeBlock),
-}
-
-/// The mempool category of ABCI requests.
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub enum MempoolRequest {
-    #[doc = include_str!("doc/request-checktx.md")]
-    CheckTx(CheckTx),
-}
-
-/// The info category of ABCI requests.
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub enum InfoRequest {
-    #[doc = include_str!("doc/request-info.md")]
-    Info(Info),
-    #[doc = include_str!("doc/request-query.md")]
-    Query(Query),
-    #[doc = include_str!("doc/request-echo.md")]
-    Echo(Echo),
-    #[doc = include_str!("doc/request-setoption.md")]
-    SetOption(SetOption),
-}
-
-/// The snapshot category of ABCI requests.
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub enum SnapshotRequest {
-    #[doc = include_str!("doc/request-listsnapshots.md")]
-    ListSnapshots,
-    #[doc = include_str!("doc/request-offersnapshot.md")]
-    OfferSnapshot(OfferSnapshot),
-    #[doc = include_str!("doc/request-loadsnapshotchunk.md")]
-    LoadSnapshotChunk(LoadSnapshotChunk),
-    #[doc = include_str!("doc/request-applysnapshotchunk.md")]
-    ApplySnapshotChunk(ApplySnapshotChunk),
-}
