+++ tendermint/src/v0_37/abci.rs
@@ -1,5 +1,5 @@
 pub mod request;
 pub mod response;
 
-pub use request::Request;
-pub use response::Response;
+pub use request::{ConsensusRequest, InfoRequest, MempoolRequest, Request, SnapshotRequest};
+pub use response::{ConsensusResponse, InfoResponse, MempoolResponse, Response, SnapshotResponse};
